# client-app/openapi-doc/api-docs.json
{"openapi":"3.0.1","info":{"title":"TripTrove Manager API","description":"Trip attractions manager","version":"25122024-2058"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/countries/{id}/details":{"put":{"tags":["Country"],"summary":"Update country details","operationId":"updateCountryDetail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCountryDetailsRequest"}}},"required":true},"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"204":{"description":"Country details are updated"}}}},"/countries/{id}/continent":{"put":{"tags":["Country"],"summary":"Update the continent name of the country","operationId":"updateCountryContinent","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCountryContinentRequest"}}},"required":true},"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"204":{"description":"Country details are updated"}}}},"/continents/{name}":{"get":{"tags":["Continent"],"summary":"Retrieve continent by name","operationId":"getContinent","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"Requested continent","content":{"*/*":{"schema":{"$ref":"#/components/schemas/GetContinentResponse"}}}}}},"put":{"tags":["Continent"],"summary":"Update continent name","operationId":"updateContinent","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateContinentRequest"}}},"required":true},"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"204":{"description":"Continent name is updated"}}},"delete":{"tags":["Continent"],"summary":"Delete continent by name","operationId":"deleteContinent","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"204":{"description":"Deleted continent by name"}}}},"/countries":{"get":{"tags":["Country"],"summary":"List paginable countries, sorted by their last updated time. If the country was never updated, sort by the creation time. Order by the given sort direction, or ascending if none is provided.","operationId":"getAllCountries","parameters":[{"name":"sd","in":"query","description":"Direction of ordering countries using last updated time, or by creation time if not updated.","required":false,"schema":{"type":"string","default":"DESC","enum":["ASC","DESC"]}},{"name":"after","in":"query","description":"Last country retrieved on the previous page. Leave empty if this is the first page.","required":true,"schema":{"$ref":"#/components/schemas/CountryParameter"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetCountryResponse"}}}}}}},"post":{"tags":["Country"],"summary":"Save new country","operationId":"saveCountry","parameters":[{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveCountryRequest"}}},"required":true},"responses":{"404":{"description":"Not Found"},"409":{"description":"Country already exists"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"201":{"description":"Country saved successfully"}}}},"/continents":{"get":{"tags":["Continent"],"summary":"List all saved continents, sorted by the time they were last updated. If the continent was never updated, sort by the creation time. Order by the given sort direction, or ascending if none is provided.","operationId":"getAllContinents","parameters":[{"name":"sd","in":"query","required":false,"schema":{"type":"string","default":"ASC","enum":["ASC","DESC"]}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GetContinentResponse"}}}}}}},"post":{"tags":["Continent"],"summary":"Save new continent","operationId":"saveContinent","parameters":[{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveContinentRequest"}}},"required":true},"responses":{"404":{"description":"Not Found"},"409":{"description":"Continent already exists"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"201":{"description":"Continent saved successfully"}}}},"/countries/{id}":{"delete":{"tags":["Country"],"summary":"Delete country by its id","operationId":"deleteCountry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}},{"name":"x-api-version","in":"header","schema":{"type":"string","enum":["1"]}}],"responses":{"404":{"description":"Not Found"},"409":{"description":"Conflict"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}},"204":{"description":"Deleted country by its id"}}}}},"components":{"schemas":{"UpdateCountryDetailsRequest":{"type":"object","properties":{"countryName":{"type":"string"}}},"UpdateCountryContinentRequest":{"type":"object","properties":{"continentName":{"type":"string"}}},"UpdateContinentRequest":{"required":["continentName"],"type":"object","properties":{"continentName":{"maxLength":64,"minLength":0,"type":"string"}}},"SaveCountryRequest":{"required":["continentName","countryName"],"type":"object","properties":{"continentName":{"maxLength":64,"minLength":0,"type":"string"},"countryName":{"maxLength":256,"minLength":0,"type":"string"}}},"SaveContinentRequest":{"required":["continentName"],"type":"object","properties":{"continentName":{"maxLength":64,"minLength":0,"type":"string"}}},"CountryParameter":{"type":"object","properties":{"countryId":{"type":"integer","format":"int32"},"updatedOn":{"type":"string","format":"date-time"}}},"GetCountryResponse":{"type":"object","properties":{"countryId":{"type":"integer","format":"int32"},"continentName":{"type":"string"},"countryName":{"type":"string"},"changedOn":{"type":"string","format":"date-time"}}},"GetContinentResponse":{"type":"object","properties":{"continentName":{"type":"string"}}}}}}